refactor(core): change validation exception, add guard in commit transaction
