--- src/Ecss.Common/Errors/ValidationError.cs
@@ -1,13 +1,0 @@
-﻿namespace Ecss.Common.Errors;
-
-public class ValidationError
-{
-    public ValidationError(string propertyName, string errorMessage)
-    {
-        PropertyName = propertyName;
-        ErrorMessage = errorMessage;
-    }
-
-    public string PropertyName { get; }
-    public string ErrorMessage { get; }
-}
