--- src/Ecss.Common/Exceptions/DuplicateException.cs
@@ -1,13 +1,0 @@
-﻿namespace Ecss.Common.Exceptions;
-
-public class DuplicateException : Exception
-{
-    public DuplicateException(string message) : base(message)
-    {
-    }
-
-    public DuplicateException(string message, Exception innerException) : base(message, innerException)
-    {
-
-    }
-}
