+++ src/Ecss.DataAccess/DependencyInjection.cs
@@ -1,6 +1,8 @@
 ﻿using Ecss.Business.Interfaces.ExternalServices;
+using Ecss.Business.Interfaces.UnitOfWorks;
 using Ecss.DataAccess.Data;
 using Ecss.DataAccess.ExternalServices;
+using Ecss.DataAccess.UnitOfWorks;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
@@ -18,6 +20,7 @@ public static class DependencyInjection
             options.Configuration = configuration.GetConnectionString("DefaultCachingConnection"));
 
         services.AddSingleton<IDateTimeProvider, DateTimeProvider>();
+        services.AddScoped<IUnitOfWork, UnitOfWork>();
 
         return services;
     }
